Declare macro COCOA_EXTENSIONS_RUNNING_ON()
Macro is used to conditionally execute code based on the version of
macOS the framework is being executed on.

It is a macro because depending on the version of macOS SDK built
against, there are two different ways to define this condition.

When built against the 10.13 SDK, @available() is used to define this
condition. On earlier SDKs, the conventional methods
[XRSystemInformation isUsingOSX…OrLater] are used.